Consolidate 17 admin dirs + 6 MPC dirs into canonical platform structures
Admin consolidation (Phase 1.1):
- Merged 17 duplicate admin* directories into single canonical admin/ platform
- admin/go/: merged real handlers (28) and services (13+) from admin_backend
- Replaced all StatusNotImplemented stub handlers with real DB-backed implementations
- Fixed all compilation errors (go build + go vet = 0 exit code)
- Merged unique services from admin_platform (oauth, payment_gateway, rate_limiter,
  chain_management, listing_management, bulk_operations, websocket_notifications,
  ticket_system) and admin_panel (cloudflare, datadog, pagerduty, backup integrations)
- Merged unique web frontend pages (CryptoCards, P2PMerchant, Features, Liquidity,
  MarginTrading) and ThemeContext/useTheme into admin/web/
- Merged admin_flutter into admin/flutter/, admin_cpp_core into admin/cpp/,
  admin_desktop into admin/desktop/, admin_extension into admin/extensions/,
  admin_ios into admin/ios/, admin_rust into admin/rust/
- Consolidated Android screens from admin_android, admin_console, admin_services,
  admin_system into admin/android/

MPC consolidation (Phase 1.2):
- Merged 6 MPC dirs into 3 platform cores: rust/mpc/, go/mpc/, cpp/mpc/
- rust/mpc/: merged mpc_wallet (Feldman-VSS, key_gen, signing, key_sharing, wallet)
  with unique Shamir SSSS from mpc/ and BIP-39 mnemonic from wallet_core_mpc
- Fixed real crypto bug: double-hashing asymmetry in sign/verify (sign was hashing
  message twice due to k256 internal RFC6979, verify only hashed once)
- cargo check passes (0 errors), cargo test --lib passes (10/10 tests)
- go/mpc/: merged enterprise_mpc Go threshold signing
- cpp/mpc/: merged C++ enterprise MPC header

+++ admin/go/internal/database/postgres.go
@@ -21,8 +21,8 @@ func Initialize(cfg *config.Config) error {
 		return fmt.Errorf("failed to parse database config: %w", err)
 	}
 
-	poolConfig.MaxConns = cfg.DatabaseMaxConns
-	poolConfig.MinConns = cfg.DatabaseMinConns
+	poolConfig.MaxConns = int32(cfg.DatabaseMaxConns)
+	poolConfig.MinConns = int32(cfg.DatabaseMinConns)
 
 	Pool, err = pgxpool.NewWithConfig(ctx, poolConfig)
 	if err != nil {
